+++ src/resources/lang/es/messages.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+	'drop_files_or_click_to_upload' => 'Suelte los archivos aquí o haga click para cargar.',
+	'media_successfully_uploaded' => 'Archivo subido correctamente',
+	'media_successfully_deleted' => 'Archivo eliminado correctamente',
+	'media_successfully_reordered' => 'Archivos reordenados correctamente',
+	'deletion_failed' => 'La eliminación ha fallado',
+	'ordering_failed' => 'El ordenamiento ha fallado',
+];